Admin/requirements.txt
# FREE VERSION - No cloud costs!
# Local AI models instead of AWS Bedrock

# Core dependencies - Updated for compatibility
pypdf
langchain
langchain-community
langchain-core
langchain-text-splitters
streamlit
python-dotenv

# FREE embeddings (replaces AWS Titan)
sentence-transformers

# FREE LLM (replaces AWS Nova)
ollama

# Vector store
faiss-cpu

# Dependencies for compatibility
httpx
tokenizers

# Optional: Better performance
torch

# Note: boto3 removed - no AWS needed!
